Feature/be/hien ht/authentication 04

+++ Infrastructure/Contexts/ServiceInCombo.cs
@@ -7,11 +7,11 @@ namespace Infrastructure.Contexts
 {
     public partial class ServiceInCombo
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public long ServiceComboId { get; set; }
-        public long ServiceId { get; set; }
+        public int ServiceComboId { get; set; }
+        public int ServiceId { get; set; }
         public int Quantity { get; set; }
 
         public virtual Service Service { get; set; }
